requirements.txt
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
coverage==5.5
docopt==0.6.2
Faker==8.13.2
fastapi==0.68.1
Flask==2.0.1
Flask-Cors==3.0.10
idna==3.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mongomock==3.23.0
packaging==21.0
pipreqs==0.4.10
pluggy==1.0.0
py==1.10.0
pydantic==1.8.2
pymongo==3.12.0
pyparsing==2.4.7
pytest==6.2.5
pytest-cov==2.12.1
python-dateutil==2.8.2
python-dotenv==0.19.0
requests==2.26.0
sentinels==1.0.0
six==1.16.0
starlette==0.14.2
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.6
Werkzeug==2.0.1
yarg==0.1.9
